Adding corrections for the lesson 07

+++ L07/src/main/java/ru.otus/l07/Nominals.java
@@ -2,13 +2,13 @@ package ru.otus.l07;
 
 enum Nominals {
     n10(10), n50(50), n100(100), n200(200), n1000(1000), n2000(2000), n5000(5000);
-    private short value;
+    private int value;
 
     Nominals(int value) {
-        this.value = (short) value;
+        this.value = value;
     }
 
-    public short getValue() {
+    public int getValue() {
         return value;
     }
